% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CAM.R
\docType{package}
\name{CAM-package}
\alias{CAM-package}
\title{Continuous Admixture Modeling (CAM)}
\description{
\pkg{CAM} includes functions to do Continuous Admixture Modeling (CAM), generate summary plots, select the best-fit model(s), generate statistics to test if the results are credible and miscellaneous functionalities.
}
\details{
See \href{https://github.com/david940408/CAM/blob/master/inst/doc/intro.md}{An Introduction to CAM package} or intro.html in under inst/doc/ subdirectory of the package for an introduction. This file demonstrates how to use the functions.
}

