% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CAMer.R
\name{CAM}
\alias{CAM}
\title{Continuous Admixture Modeling (CAM)}
\usage{
CAM(rawld, m1, T = 500L, isolation = TRUE, fast.search = TRUE,
  max.duration = 150L, LD.parallel = TRUE, LD.clusternum,
  single.parallel = isolation && !fast.search, single.clusternum = 4L)
}
\arguments{
\item{rawld}{a string representing the path of the .rawld file or a data frame read from the .rawld file by \code{read.table}.}

\item{m1}{the admixture proportion of population 1. If m2 is the admixing proportion of population 2, then m1+m2=1.}

\item{T}{the most ancient generation to be searched. Defaults to 500.}

\item{isolation}{\code{TRUE} if the models used for fitting are HI, CGF1-I, CGF2-I and GA-I; \code{FALSE} if the models used for fitting are HI, CGF1, CGF2 and GA. Defaults to \code{TRUE}.}

\item{fast.search}{only used when \code{isolation=TRUE}. \code{TRUE} to use the fast searching algorithm, which sometimes gives slightly wider time intervals than the slow searching algorithm. Defaults to \code{TRUE}.}

\item{max.duration}{Defaults to 150. See "Details".}

\item{LD.parallel}{a logical expression indicating whether each LD decay curve should be computed parallely. Defaults to \code{TRUE}.}

\item{LD.clusternum}{the number of clusters in parallel computation. See "Details".}

\item{single.parallel}{a logical expression. See "Details".}

\item{single.clusternum}{the number of clusters in parallel computation. Defaults to 4 for the four models. Used if \code{single.parallel=TRUE}.}
}
\value{
an object of S3 class "CAM". A list \code{CAM.list} consisting of some basic information of function call, N objects of "CAM.single" class (where N is the number of LD decay curves in the .rawld file), the fitted Ld decay curve fitted by the previous method (up to some truncation according to \code{max.index}) and a summary table containing the parameter estimates for each model and each curve, and diagnostic statistics msE and quasi-F. For details of "CAM.single" class, see \code{\link{singleCAM}}.

There is a special method of \code{plot} and \code{print} for this class.
}
\description{
Estimate admixture time intervals/points for HI, CGF1(-I), CGF2(-I) and GA(-I) respectively for all Ld decay curves in a .rawld file.
}
\details{
\code{max.duration} is only used when \code{isolation=TRUE} and \code{fast.search=FALSE}. The maximal duration of admixture \eqn{n} to be considered as possible. Smaller values can make the slow searching algorithm faster. If \code{max.duration>T}, it will be set to be \code{T}.

\code{LD.clusternum} is used if \code{LD.parallel=TRUE}. If not specified, it is set to be the number of LD decay curves in the .rawld file.

\code{single.parallel} indicates whether parallel computation should be used when computing a single LD decay curve. Defaults to \code{TRUE} if \code{isolation=TRUE,fast.search=FALSE} and \code{FALSE} otherwise.

The .rawld file should include exactly one column nameed "Distance" in Morgan, exactly one column named "Combined_LD", several columns named "Jack?" representing Jackknives where ? is a number and exactly one column named "Fitted" representing the fitted LD decay curve using the previous method. This function fits "Combined_LD" and all Jackknives using all models. See \code{\link{singleCAM}} for further details of fitting algorithm for each admixture induced LD (ALD) decay curve decay curve.

If the last entry of Distence in the .rawld file is greater than 10, a warning of unit will be given.

If the estimated time intervals/points cover \code{T}, a warning of too small \code{T} is given. The user should re-run the function with a larger \code{T} so that optimal time intervals/points can be reached.

Require \pkg{parallel} or \pkg{snow} package installed if \code{LD.parallel=TRUE} or \code{single.parallel=TRUE}. For newer versions of \code{R (>=2.14.0)}, \pkg{parallel} is in R-core. If only \pkg{snow} is available, it is recommended to library it before using the parallel computing funcationality. When only \pkg{snow} is available, it will be \code{require}-d and hence the search path will be changed; if \pkg{parallel} is available, it will be used but the search path will not be changed. One may go to \url{https://cran.r-project.org/src/contrib/Archive/snow/} to download and install older versions of \pkg{snow} if the version of \code{R} is too old. If neither of the packages is available but \code{LD.parallel=TRUE} or \code{single.parallel=TRUE}, the function will compute sequentially with messages.

Be aware that when the computational cost is small (e.g. \code{isolation=FALSE} or \code{T=20L,isoaltion=TRUE,fast.search=FALSE,max.duration=10L}), using parallel computation for single LD decay curves can result in longer computation time.
}
\note{
When \code{LD.parallel=TRUE} or \code{single.parallel=TRUE}, it is not recommended to terminate the execution of the function. If \pkg{parallel} package is available, it is said that \code{\link[parallel]{setDefaultCluster}} from \pkg{parallel} can be used to remove the registered cluster, but real experiments do not support this; fortunately, these unused clusters will be removed automatically later, but with warnings. If only \pkg{snow} package is available, according to \url{http://homepage.stat.uiowa.edu/~luke/R/cluster/cluster.html}, "don't interrupt a snow computation". The ultimate method to close the unused clusters is probably to quit the R session.

Do care about memory allocation, especially when both \code{LD.parallel=TRUE} and \code{single.parallel=TRUE}.

It is possible that this function opens several nodes but non of them is computing, and hence the execution does not stop, especially when both \code{LD.parallel=TRUE} and \code{single.parallel=TRUE}. The cause has not been identified yet. The current solution is to terminate the function by hand and re-run the function with fewer cores (e.g. set \code{single.parallel=FALSE}).
}
\examples{
data(GA_I)

#fit models with isolation=FALSE.
fit<-CAM(GA_I,m1=0.3,T=150L,isolation=FALSE)
fit
\dontrun{
plot(fit) #may not be able to display
}
#Bad fitting indicates isolation=TRUE should be tried.
fit<-CAM(GA_I,m1=0.3,T=150L,isolation=TRUE)
fit
fit$summary
\dontrun{
plot(fit) #may not be able to display
plot(fit,"plot.pdf") #plot to a .pdf file
}

data(CGF_50)
fit<-CAM(CGF_50,0.3,20L,isolation=FALSE,LD.parallel=FALSE) #with warnings
fit

\dontrun{
#passing a file path to the argument `rawld=`
fit<-CAM("CGF_50.rawld",0.3)
}
}
\seealso{
\code{\link{construct.CAM}}, \code{\link{reconstruct.fitted}}, \code{\link{conclude.model}}
}

