% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CAMer.R
\docType{package}
\name{CAMer}
\alias{CAMer}
\alias{CAMer-package}
\title{Continuous Admixture Modeler (CAMer)}
\description{
\pkg{CAMer} includes functions to do Continuous Admixture Modeling (CAM), generate summary plots, select the best-fit model(s), generate statistics to test if the results are credible and miscellaneous functionalities.
}
\details{
See \href{https://github.com/david940408/CAMer/blob/master/inst/doc/intro.md}{An Introduction to CAMer package} or intro.html in under inst/doc/ subdirectory of the package for an introduction. This file demonstrates how to use the functions.
}
\references{
Zhou Y, Qiu H, Xu S (2015). _Modeling Continuous Admixture_. (under review)
}

