% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CAM.R
\name{GA_I}
\alias{GA_I}
\title{Simulated .rawld File for GA-I}
\format{A data frame with 3497 rows and 21 variables}
\description{
A data frame read from a .rawld file by \code{read.table}. Forward simulation data using 10 chromosomes where the true model is GA-I (start from the 100th generation and end at the 30th generation) with admixture proportion of population 1 being 0.3.
}
\details{
The useful variables are as follows:
\itemize{
\item Distance: genetic distance
\item Combined_LD: weighted LD
\item Fitted: the fitted LD decay curve using the previous method
\item Jack?: Jackknives (?=1,2,...)
}
}

