% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CAMer.R
\name{HI}
\alias{HI}
\title{Time Inference of HI model}
\usage{
HI(rawld, m1, T = 500L, LD.parallel = TRUE, LD.clusternum)
}
\arguments{
\item{rawld}{a string representing the path of the .rawld file or a data frame read from the .rawld file by \code{read.table}.}

\item{m1}{the admixture proportion of population 1. If m2 is the admixing proportion of population 2, then m1+m2=1.}

\item{T}{the most ancient generation to be searched. Defaults to 500.}

\item{LD.parallel}{a logical expression indicating whether each LD decay curve should be computed parallely. Defaults to \code{TRUE}.}

\item{LD.clusternum}{the number of clusters in parallel computation. See "Details".}
}
\value{
an object of S3 class "CAM". A list \code{CAM.list} consisting of some basic information of function call, N objects of "CAM.single" class (where N is the number of LD decay curves in the .rawld file) calculated by \code{\link{singleHI}}, the fitted Ld decay curve fitted by the previous method (up to some truncation according to \code{max.index}) and a summary table containing the parameter estimates for each model and each curve, and diagnostic statistics msE and quasi-F. See \code{\link{singleCAM}} for more details of this kind of "CAM.songle" class.

There is a special method of \code{print} for this class.
}
\description{
Estimate admixture time intervals/points for HI, CGF1(-I), CGF2(-I) and GA(-I) respectively for all Ld decay curves in a .rawld file.
}
\details{
\code{LD.clusternum} is used if \code{LD.parallel=TRUE}. If not specified, it is set to be the number of LD decay curves in the .rawld file.

The .rawld file should include exactly one column nameed "Distance" in Morgan, exactly one column named "Combined_LD", several columns named "Jack?" representing Jackknives where ? is a number and exactly one column named "Fitted" representing the fitted LD decay curve using the previous method. This function fits "Combined_LD" and all Jackknives using HI model. See \code{\link{singleHI}} for further details of fitting algorithm for each LD decay curve.

If the last entry of Distence in the .rawld file is greater than 10, a warning of unit will be given.

If the estimated time intervals/points cover \code{T}, a warning of too small \code{T} is given. The user should re-run the function with a larger \code{T} so that optimal time intervals/points can be reached.

Require \pkg{parallel} or \pkg{snow} package installed if \code{LD.parallel=TRUE} or \code{single.parallel=TRUE}. For newer versions of \code{R (>=2.14.0)}, \pkg{parallel} is in R-core. If only \pkg{snow} is available, it is recommended to library it before using the parallel computing funcationality. When only \pkg{snow} is available, it will be \code{require}-d and hence the search path will be changed; if \pkg{parallel} is available, it will be used but the search path will not be changed. One may go to \url{https://cran.r-project.org/src/contrib/Archive/snow/} to download and install older versions of \pkg{snow} if the \code{R} version is too old. If neither of the packages is available but \code{LD.parallel=TRUE} or \code{single.parallel=TRUE}, the function will compute sequentially with messages.

Be aware that when the computational cost is small (e.g. \code{T=20L}), using parallel computation for single LD decay curves can result in longer computation time.
}
\note{
Although the output is a "CAM" class object, it should \emph{NOT} be passed to \code{\link{plot.CAM}}. Its summary table should \emph{NOT} be passed to \code{\link{construct.CAM}} either.

When \code{LD.parallel=TRUE}, it is not recommended to terminate the execution of the function. If \pkg{parallel} package is available, it is said that \code{\link[parallel]{setDefaultCluster}} from \pkg{parallel} can be used to remove the registered cluster, but real experiments do not support this; fortunately, these unused clusters will be removed automatically later, but with warnings. If only \pkg{snow} package is available, according to \url{http://homepage.stat.uiowa.edu/~luke/R/cluster/cluster.html}, "don't interrupt a snow computation". The ultimate method to close the unused clusters is probably to quit the R session.
}
\examples{
data(GA_I)
fit<-HI(GA_I,m1=0.3,T=150L)
fit

\dontrun{
#passing a file path to the argument `rawld=`
fit<-CAM("CGF_50.rawld",m1=0.3,T=150L)
}
}
\seealso{
\code{\link{singleHI}}, \code{\link{CAM}}
}

