% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CAM.R
\name{plot.CAM}
\alias{plot.CAM}
\title{Summary Plots for "CAM" Class}
\usage{
\method{plot}{CAM}(x, filename, model.cols = matrix(c("#ffa1c2B2",
  "#b50d37B2", "#da577c", "#9bff94", "#0fbd02", "#0fbd02", "#e9a1ff", "#9111b8",
  "#9111b8", "#7af8ff", "#1ea1a8", "#1ea1a8"), ncol = 4), box.log = TRUE,
  alpha = 0.6, fit.lwd = 3, LD.col = "black", LD.lwd = 1, ...)
}
\arguments{
\item{x}{an object of "CAM" class}

\item{filename}{pdf file path.}

\item{model.cols}{a matrix of colors. See "Details"}

\item{box.log}{a logical expression. If \code{TRUE}, the scale of the axis will be in log scale. Defaults to \code{TRUE}}

\item{alpha}{a numeric value in [0,1] representitng alpha of the fitted LD decay curves for Combined LD. Defaults to 0.6. Can be a vector representing different alphas for different models. Ignored if alpha is specified in the third row of \code{model.cols}.}

\item{fit.lwd}{line width of the fitted LD decay curve for Combined LD of the four models. Defaults to 3. Can be a vector representing different line widths for different models.}

\item{LD.col}{color of the original Combined LD curve. Defaults to "black".}

\item{LD.lwd}{line width of the original Combined LD curve. Defaults to 1.}

\item{...}{other graphical arguments passed to basic functions like \code{plot}.}
}
\description{
Summary Plots for "CAM" Class
}
\details{
The function generates three plots in a device. The plot on the top left is the estimated time intervals/points for the four models. The color depth of segments/points corresponds to how many intervals/points covers this part in Jackknives. The deeper the color, the more estimates from Jackknives cover this part. The plot on the top right is the boxplot of msE for the four models. The third plot shows the fitting of four models to \code{Combined_LD} in the .rawld file. The numbers after model names in the legend are quasi-F values of the four models for \code{Combined_LD}.

If \code{filename} is set, plot to the .pdf file, otherwise plot to the current device. The function is specially designed for a .pdf plot with width being 9.6 and height being 7.2. To add things to the plot and then save it to a file, better to set the size as above. May not be able to plot directly in an R window.

The colors in Column 1/2/3/4 correspond to the colors for HI/CGF1(-I)/CGF2(-I)/GA(-I). The first/second row of \code{model.cols} is the lightest/deepest possible color in the "Time Intervals/Points" plot. The third row of \code{model.cols} is the color for "msE Boxplot" and "Fitting of Models" plot. The colors will be converted to RGB colors by \code{col2rgb}.
}
\note{
It is not recommended to pass other arguments to basic functions like \code{plot} in \code{...}.
}
\examples{
\dontrun{
data(CGF_50)
fit<-CAM(CGF_50,0.3,10L,isolation=FALSE,LD.parallel=FALSE)
plot(fit,"bad_fitting.pdf")

#may not be able to display
plot(fit,model.cols=matrix(c("pink","red","pink",
                             "lightseagreen","green","green",
                             "skyblue","blue","blue",
                             "yellow","orange","orange"),ncol=4),
     box.log=FALSE,alpha=1,fit.lwd=1,LD.col="gray",LD.lwd=3)
}
}
\seealso{
\code{\link{CAM}}, \code{\link[grDevices]{rgb}}, \code{\link[grDevices]{col2rgb}}
}

