% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CAM.R
\name{print.CAM.conclusion}
\alias{print.CAM.conclusion}
\title{Print Method for "CAM.conclusion" Class}
\usage{
\method{print}{CAM.conclusion}(x)
}
\arguments{
\item{x}{an object of "CAM.conclusion" class}
}
\description{
Print Method for "CAM.conclusion" Class
}
\seealso{
\code{\link{conclude.model}}
}

