% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CAM.R
\name{print.CAM}
\alias{print.CAM}
\title{Print Method for "CAM" Class}
\usage{
\method{print}{CAM}(x, ...)
}
\arguments{
\item{x}{a "CAM" class object}

\item{...}{further arguments}
}
\description{
Print Method for "CAM" Class
}
\details{
The print method for "CAM" Class. Include:
\itemize{
\item Function call (if available from the object)
\item Total length of LD decay curve
\item a data frame containing the estimated intervals/points, msE and quasi-F for each model and each LD decay curve. The time point for HI model is stored in \code{Start} variable.
}
}
\seealso{
\code{\link{CAM}}
}

