% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CAM.R
\name{print.CAM.single}
\alias{print.CAM.single}
\title{Print Method for "CAM.single" Class}
\usage{
\method{print}{CAM.single}(x, ...)
}
\arguments{
\item{x}{a "CAM.single" class object}

\item{...}{further arguments}
}
\description{
Print Method for "CAM.single" Class
}
\details{
Print a very brief summary of a "CAM.single" class object. Include:
\itemize{
\item Function call
\item length of used LD (excluding first few values if necessary)
\item a data frame containing the estimated time intervals/points and corresponding msE. The time point for HI model is stored in \code{Start} variable.
}
}
\seealso{
\code{\link{singleCAM}}
}

