% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CAM.R
\name{reconstruct.fitted}
\alias{reconstruct.fitted}
\title{Reconstruct the Fitted LD Decay Curves of Four Models}
\usage{
reconstruct.fitted(CAM.single)
}
\arguments{
\item{CAM.single}{an object of "CAM.single" class}
}
\value{
a list consisting of the four fitted curves
}
\description{
Given an object of "CAM.single" class, reconstruct the fitted LD decay curves of the four models (from \code{\link{singleCAM}}) or HI model (from \code{\link{singleHI}}).
}
\examples{
data(CGF_50)
Z<-CGF_50$Combined_LD
d<-CGF_50$Distance

fit<-singleCAM(d,Z,m1=0.3,T=100L,isolation=FALSE)
fitted.curves<-reconstruct.fitted(fit)

fit<-singleHI(d,Z,m1=0.3,T=100L)
fitted.curves<-reconstruct.fitted(fit)
}
\seealso{
\code{\link{singleCAM}}, \code{\link{singleHI}} \code{\link{construct.CAM}}
}

