% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CAMer.R
\name{singleHI}
\alias{singleHI}
\title{Time Inference of HI model for a Single LD Decay Curve}
\usage{
singleHI(d, Z, m1, T = 500L)
}
\arguments{
\item{d}{the numeric vector of genetic distance (Morgan) of LD decay curve}

\item{Z}{the numeric vector of LD decay curve}

\item{m1}{the admixture proportion of population 1. If m2 is the admixing proportion of population 2, then m1+m2=1.}

\item{T}{the most ancient generation to be searched. Defaults to 500.}
}
\value{
an object of S3 class "CAM.single". A list consisting of:
\item{call}{the matched call}
\item{maxindex}{the index of the maximal value in \code{Z} See "Details".}
\item{d,Z}{identical to function inputs up to some truncation. See "Details"}
\item{T,isolation}{identical to function inputs}
\item{A}{numeric matrix \eqn{A} with the \eqn{(i,j)}-th entry being \eqn{\text{exp}(-j \cdot d_i)}, \eqn{d_i} meaning the \eqn{i}-th entry of \code{d} and \eqn{j} meaning the genertion.}
\item{m1,m2}{admixture proportion of population 1 and 2}
\item{estimate}{a list of estimates. Each element contains the estimated parameters \eqn{m}, \eqn{n}, \eqn{\theta_0}, \eqn{\theta_1}, starting generation, ending generation and the corresponding ssE and msE. The time point for HI model is stored in \code{start} variable.}
\item{summary}{a data frame containing the information in \code{estimate} in a compact form}
}
\description{
Find the estimated time point for HI model and corresponding statictis (ssE, msE, etc.) for a single LD decay curve (e.g. Combined_LD or Jack? in a .rawld file).
}
\details{
This function is similar to \code{\link{singleCAM}}, except that it only considers HI model as the core model.
}
\examples{
data(CGF_50)
d<-CGF_50$Distance
Z<-CGF_50$Combined_LD
fitHI<-singleHI(d,Z,m1=.3,T=70L)
fitHI
}
\seealso{
\code{\link{singleCAM}}, \code{\link{HI}}, \code{\link{reconstruct.fitted}}
}

