% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CAM.R
\docType{package}
\name{CAM-package}
\alias{CAM-package}
\title{Continuous Admixture Modeling (CAM)}
\description{
\pkg{CAM} includes functions to do Continuous Admixture Modeling (CAM), generate summary plots, select the best-fit model(s), generate statistics to test if the results are credible and miscellaneous functionalities.
}
\details{
Type \code{browseVignettes(CAM)} in R for the vignettes to see some examples of how to use the functions.
}

