% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAMer.R
\name{CGF_50}
\alias{CGF_50}
\title{Simulated .rawld File for CGF1}
\format{A data frame with 3497 rows and 21 variables}
\description{
A data frame read from a .rawld file by \code{read.table}. Forward simulation data using 10 chromosomes where the true model is CGF1 (CGF with population 1 being the recipient) (50 generations) with admixture proportion of population 1 being 0.3.
}
\details{
The useful variables are as follows:
\itemize{
\item Distance: genetic distance
\item Combined_LD: weighted LD
\item Fitted: the fitted LD decay curve using the previous method
\item Jack?: Jackknives (?=1,2,...)
}
}

